ubuntu_server/requirements.txt
pyautogui
opencv-python
numpy
PyOpenGL
PyOpenGL_accelerate
python-xlib
glfw